requirements.txt
opencv-python-headless
numpy
requests
python-dotenv
tensorflow-lite-runtime
